ssp_api_new/requirements.txt
flask==3.0.0
python-oracledb==2.0.0
python-gitlab==3.2.0
jinja2==3.1.2
loguru==0.7.0
marshmallow==3.19.0
python-dotenv==1.0.0
pyjwt==2.8.0